chore rename files to match page vs components
