+++ Singer.API/ClientApp/package.json
@@ -1,6 +1,6 @@
 {
    "name": "Singer",
-   "version": "0.1.0-issue-AB-23-user-profile-page.1333",
+   "version": "0.1.0-issue-AB-23-user-profile-page.1334",
    "scripts": {
       "ng": "ng",
       "start": "ng serve",
