+++ Singer.API/ClientApp/src/app/modules/shared/components/account-info/account-info-base/account-info-page.component.spec.ts
@@ -1,19 +1,18 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { AccountInfoPageComponent } from './account-info-page.component';
 
-import { AccountInfoBaseComponent } from './account-info-base.component';
-
-describe('AccountInfoBaseComponent', () => {
-   let component: AccountInfoBaseComponent;
-   let fixture: ComponentFixture<AccountInfoBaseComponent>;
+describe('AccountInfoPageComponent', () => {
+   let component: AccountInfoPageComponent;
+   let fixture: ComponentFixture<AccountInfoPageComponent>;
 
    beforeEach(async(() => {
       TestBed.configureTestingModule({
-         declarations: [AccountInfoBaseComponent],
+         declarations: [AccountInfoPageComponent],
       }).compileComponents();
    }));
 
    beforeEach(() => {
-      fixture = TestBed.createComponent(AccountInfoBaseComponent);
+      fixture = TestBed.createComponent(AccountInfoPageComponent);
       component = fixture.componentInstance;
       fixture.detectChanges();
    });
