+++ Singer.API/ClientApp/src/app/modules/shared/components/account-info/account-info-summary/account-info-summary.component.html
